+++ lib/src/config/model/theme_config.dart
@@ -1,42 +1,40 @@
+import 'package:flutter/painting.dart';
 import 'package:vuitv/vuitv.dart';
+
+const _kDefaultPrimaryColor = Color(0xFFe21d55);
 
 ///Theme create by config
 class ThemeConfig {
   ///Theme create by config
   const ThemeConfig({
-    this.mainColor = 'e21d55',
-    this.primaryColor = '',
-    this.secondaryColor = '',
-    this.backgroundColor = '',
-    this.textColor = '',
+    this.primaryColor = _kDefaultPrimaryColor,
+    this.secondaryColor,
+    this.backgroundColor,
+    this.textColor,
     this.fontFamily = '',
     this.fontHeader = '',
   });
 
   ///Theme fromJson
   ThemeConfig.fromJson(Map<String, dynamic> json)
-      : mainColor = json.getString('mainColor') ?? '',
-        primaryColor = json.getString('primaryColor') ?? '',
-        secondaryColor = json.getString('secondaryColor') ?? '',
-        backgroundColor = json.getString('backgroundColor') ?? '',
-        textColor = json.getString('textColor') ?? '',
+      : primaryColor = HexColor.fromJson(json.getString('primaryColor')) ?? _kDefaultPrimaryColor,
+        secondaryColor = HexColor.fromJson(json.getString('secondaryColor')),
+        backgroundColor = HexColor.fromJson(json.getString('backgroundColor')),
+        textColor = HexColor.fromJson(json.getString('textColor')),
         fontFamily = json.getString('fontFamily') ?? '',
         fontHeader = json.getString('fontHeader') ?? '';
 
   ///
-  final String mainColor;
+  final Color primaryColor;
 
   ///
-  final String primaryColor;
+  final Color? secondaryColor;
 
   ///
-  final String secondaryColor;
+  final Color? backgroundColor;
 
   ///
-  final String backgroundColor;
-
-  ///
-  final String textColor;
+  final Color? textColor;
 
   ///
   final String fontFamily;
@@ -47,11 +45,10 @@ class ThemeConfig {
   ///
   Map<String, dynamic> toJson() {
     final map = <String, dynamic>{};
-    map['mainColor'] = mainColor;
-    map['primaryColor'] = primaryColor;
-    map['secondaryColor'] = secondaryColor;
-    map['backgroundColor'] = backgroundColor;
-    map['textColor'] = textColor;
+    map['primaryColor'] = primaryColor.toJson();
+    map['secondaryColor'] = secondaryColor?.toJson();
+    map['backgroundColor'] = backgroundColor?.toJson();
+    map['textColor'] = textColor?.toJson();
     map.removeWhere((key, dynamic value) => value == null);
     return map;
   }